requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
pathspec==0.6.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
Random-Word==1.0.4
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
